requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
Flox-lib>=0.20.0
pyperclip>=1.8.0